(15-1) RED: 서로 다른 화폐 더하기,컴파일 실패하는 테스트

+++ src/main/kotlin/tdd/money/Money.kt
@@ -3,7 +3,7 @@ package tdd.money
 class Money(val amount: Double, val currency: String) : CurrencyExpression {
     constructor(amount: Number, currency: String) : this(amount.toDouble(), currency)
 
-    fun times(multiplier: Double): Money = Money(amount * multiplier, this.currency)
+    fun times(multiplier: Int): Money = Money(amount * multiplier, this.currency)
 
     operator fun plus(addend: Money): Sum = Sum(this, addend)
     override fun reduce(bank: Bank, to: String): Money {
